% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_extract_param.R
\name{afm_extract_param}
\alias{afm_extract_param}
\title{Extract parameter values from the AFM header}
\usage{
afm_extract_param(header, parameter)
}
\arguments{
\item{header}{The AFM file header as a character vector.}

\item{parameter}{The desired parameter. If you are trying to extract a
parameter from the header, you must use a character vector that matches
exactly (case sensitive) the parameter in the header. ie. "Scan Rate"}
}
\value{
If the parameter was found in the header, it will be returned either
  as numeric or character. If the parameter was not found in the header, it
  will be returned as supplied and a warning will be triggered
}
\description{
This is a flexible function that will first try to extract the supplied
parameter from the header. If the parameter is not found in the header, it
will simply return the parameter value that you specified. This allows for
addition of parameters that are not included in the nanoscope output, but
might be important for a user such as "treatment" "cell#" "material" etc.
}
