% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_extract_scan_points.R
\name{afm_extract_scan_points}
\alias{afm_extract_scan_points}
\title{Format scan data as a tibble}
\usage{
afm_extract_scan_points(data)
}
\arguments{
\item{data}{The afm scan data (header removed) as a character vector}
}
\value{
A tibble with each column as one channel
}
\description{
This function takes a character vector (usually from readLines()) of afm scan
data and formats it as a nice tibble where each column is one afm channel
(ie. height, modulus)
}
