% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_trim.R
\name{afm_hist}
\alias{afm_hist}
\title{Plot histogram of afm map}
\usage{
afm_hist(data = NULL, channel = NULL)
}
\arguments{
\item{data}{an "afm_scan" data object}

\item{channel}{the desired channel column name (eg. "Height_Sensor(nm)")}
}
\value{
a histogram of values for the selected channel
}
\description{
This is a wrapper to quickly plot a histogram of values in an afm map. To
make a customizable histogram with more options you should input your
scan_data to ggplot2 and design your own plot.
}
\examples{
\donttest{load("data/half_um_scan.rda")
afm_hist(half_um_scan, channel = "Height_Sensor(nm)")}
}
