% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_matrix.R
\name{afm_matrix}
\alias{afm_matrix}
\title{Takes a vector of afm scan data and turns it into a matrix}
\usage{
afm_matrix(data, samps_line, afm_lines, channel = NULL)
}
\arguments{
\item{data}{A tibble containing the scan data for several channels}

\item{samps_line}{samples per line for the scan}

\item{afm_lines}{number of lines in the scan}

\item{channel}{the channel you want to turn into a matrix}
}
\value{
a matrix with dimensions [lines, samples_per_line]
}
\description{
Takes a vector of afm scan data and turns it into a matrix
}
