% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_scan.R
\name{afm_scan}
\alias{afm_scan}
\title{Constructs an "afm_scan" S3 object from inputs}
\usage{
afm_scan(scan_points = NULL, scanSize = NULL, sampsPerLine = NULL,
  afmLines = NULL, maps = list(), optional_params = list())
}
\arguments{
\item{scan_points}{(required) A tibble with the scan data}

\item{scanSize}{(required) The scan size}

\item{sampsPerLine}{(required) The samples per line}

\item{afmLines}{(required) The number of lines}

\item{maps}{(optional) afm maps (matrices)}

\item{optional_params}{(optional) Any other parameters to include (see
details)}
}
\value{
An S3 object of class "afm_scan"
}
\description{
Constructs an "afm_scan" S3 object from inputs
}
\details{
This function is typically used under the hood by afm_read_bruker,
  but it can be used to construct an afm_scan object from a list of values
  which could be useful as a control or for teaching. The first four
  parameters are required, while the maps and optional_params are optional.
  Supplied maps should be a list of matrices representing the matrix
  representation of different afm channels. Other parameters such as
  experimental details should be supplied as a list to optional_params.
}
