% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_trim.R
\name{afm_trim}
\alias{afm_trim}
\title{Trim extreme values from afm maps}
\usage{
afm_trim(data = NULL, matrix = NULL, lower = 0.025, upper = 0.975,
  min = NULL, max = NULL)
}
\arguments{
\item{data}{an "afm_scan" data object}

\item{matrix}{the name of the afm map you want to trim}

\item{lower}{the lower quantile (default 0.025)}

\item{upper}{the upper quantile (default 0.975)}

\item{min}{the minimum value cutoff (overrides quantile)}

\item{max}{the maximum value cutoff (overrides quantile)}
}
\value{
an "afm_scan" data object with the trimmed matrix added under `maps`
}
\description{
This function will floor extreme values from an afm map. You can inspect the
distribution of values in an afm map with `afm_hist()` and then trim extreme
values either by quantile or specific cutoff. By default the cutoffs will be
set to the upper 97.5% and 2.5% quantiles. You can adjust these values, or
you can set the minimum and maximum values specifically using the `min` and
`max` arguments. The trimmed data will be added to your afm_scan object under
`maps` as eg. Height_Sensor_matrix_trimmed. The trimmed data should never be
used for calculation, only to improve image quality.
}
\examples{
\donttest{load("data/half_um_scan.rda")
afm_trim(half_um_scan, matrix = "Height_Sensor_matrix")}
}
