% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_um_scan}
\alias{two_um_scan}
\title{A 2um x 2um AFM scan}
\format{An object of class "afm_data"
\describe{
  \item{params}{the critical parameters of scan size (nm), lines, and samples per line}
  \item{scan_data}{a dataframe with the raw scan data}
  \item{opt_params}{optional parameters (empty)}
  \item{maps}{the matrices of dimension [lines, samples per line] corresponding to the channels in scan_data}
  ...
}}
\usage{
two_um_scan
}
\description{
This is an AFM scan of the inner surface of the plant cell wall
collected on a Bruker Icon AFM with a scan-asyst fluid+ tip. The
data was collected in PeakForce QNM mode, and contains channels for
height, peakforce, DMT modulus, adhesion, dissipation
}
\keyword{datasets}
